@@ -20,18 +20,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^20.1.0",
+    "@angular/animations": "^20.2.0",
     "@angular/cdk": "^20.2.0",
-    "@angular/common": "^20.1.0",
-    "@angular/compiler": "^20.1.0",
-    "@angular/core": "^20.1.0",
-    "@angular/forms": "^20.1.0",
-    "@angular/language-service": "^20.1.0",
-    "@angular/localize": "^20.1.0",
+    "@angular/common": "^20.2.0",
+    "@angular/compiler": "^20.2.0",
+    "@angular/core": "^20.2.0",
+    "@angular/forms": "^20.2.0",
+    "@angular/language-service": "^20.2.0",
+    "@angular/localize": "^20.2.0",
     "@angular/material": "^20.2.0",
-    "@angular/platform-browser": "^20.1.0",
-    "@angular/platform-browser-dynamic": "^20.1.0",
-    "@angular/router": "^20.1.0",
+    "@angular/platform-browser": "^20.2.0",
+    "@angular/platform-browser-dynamic": "^20.2.0",
+    "@angular/router": "^20.2.0",
     "@coreui/angular": "^5.5.5",
     "@coreui/angular-chartjs": "~5.5.5",
     "@coreui/chartjs": "~4.1.0",
@@ -39,13 +39,12 @@
     "@coreui/icons": "^3.0.1",
     "@coreui/icons-angular": "~5.5.5",
     "@coreui/utils": "^2.0.2",
-    "@genezio-sdk/f123dashboard": "^1.0.0-prod",
+    "@f123dashboard/shared": "1.0.0",
     "@ng-bootstrap/ng-bootstrap": "^19.0.1",
     "angularx-flatpickr": "^8.0.0",
     "chart.js": "^4.5.1",
     "flag-icons": "^7.2.3",
     "flatpickr": "^4.6.13",
-    "genezio": "^2.6.3",
     "lodash-es": "^4.17.21",
     "ngx-scrollbar": "^13.0.3",
     "rxjs": "~7.8.2",
@@ -53,9 +52,9 @@
     "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular/build": "^20.1.0",
-    "@angular/cli": "^20.1.0",
-    "@angular/compiler-cli": "^20.1.0",
+    "@angular/build": "^20.2.0",
+    "@angular/cli": "^20.2.0",
+    "@angular/compiler-cli": "^20.2.0",
     "@eslint/js": "^9.18.0",
     "@types/jasmine": "^5.1.8",
     "@types/lodash-es": "^4.17.12",
@@ -75,7 +74,7 @@
     "parse5": "8.0.0"
   },
   "engines": {
-    "node": "^20.00.0 || ^22.12.0 || ^24.0.0",
-    "npm": ">= 9"
+    "node": ">=20.0.0",
+    "npm": ">=9.0.0"
   }
 }
