@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  apiBaseUrl: 'https://f123dashboard.app.genez.io'
+  apiBaseUrl: '/api'
 };
